Revert some changes for deprioritized development
