--- sources/streamer/core/ga-common-kmd-safe.h
@@ -1,32 +1,0 @@
-// Copyright (C) 2022 Intel Corporation
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions
-// and limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * @file
- * header: common GA functions and macros that can be safely included by kernal mode drivers
- */
-
-#ifndef __GA_COMMON_KMD_SAFE_H__
-#define __GA_COMMON_KMD_SAFE_H__
-
-enum Severity {
-    ERR,
-    WARNING,
-    INFO,
-    DBG
-};
-
-#endif
